docs: Add the first draft of the Introduction #6 and Literature Review #9
Include the motivation for this work and an overview of the topics in the Introduction.
Add 3 first topics of the Literature Review:
Reliability Engineering and Power Electronics
Condition Monitoring
The IGBT's functions and Failure Mechanisms

The last section still needs to explain the behavior of the IGBT's modules in the failures: Open-circuit or short-circuit.