Create the folder structure.
This work has the following folder structure:
./Dissertation -> Latex code and figures for the written document.
./figuras -> useful figures.
./source -> files for the simulation and case analysis.
./source/images -> png or jpg files from the case analysis.
./source/Simulations -> all simulation files.
./source/tools -> all the custom classes and functions for the case analysis.

The structure for add a case simulation is:
The folders in the ./source/Simulations/ have to be <model + number> to identify the preset machine model.
For example, ./source/Simulations/model15/
Inside the model folder, create a folder for each simulation of that model. For example, model15/openCircuit/ for open circuit tests.
If there are more the one setup for a model, such as different controls, the model folder have to specify this cases.
All folders inside the ./source/tools and ./source/Simulations have __init__.py to be imported in Python modules.