Fix FFT function.
Fix FFT function at ./source/tools/FFT.py to calculate the Fourier transform of the signal for real elements with the an array of signal elements and the sampling period of the signal. The method calcFFT returns an array with frequency elements and another with corresponding FFT values. The function is tested in ./source/fullSpeedScalar.py.