Begin to implement fault detection algorithm.
Add file jungChoi.py to implement Jung Choi algorithm to detect open-circuit faults of the input signal.
The file receive the time series of the simulations and part the signal into 1.0 s intervals in order to apply the algorithms. A flag is raised when fault occurs.
Fix tools/RMS.py : The loop inside the function RMS was rewritten to fix an indexing problem with the final RMS array.