Issue #10:Plots similar to the Jung-Hyun Choi paper.
Modify the jungHyunChoiAlgorithm.py to add multiple plots to the end of the file similar to the figure 13 of the paper:
'A Diagnostic Method of Simultaneous Open Switch Faults in Inverter-Fed Linear Induction Motor Drive for Reliability Enhancement' from Jyun-Hyun Choi.
Enhancements:
Every function used in the tools library is made to process an interval of a signal. The simulation is divided into small intervals and each one is processed with functions to get FFT, the first 3 harmonics, RMS value, average, fault status (0 or 1) and fault modes (0 to 4). The calculus in a single loop with a preset number of intervals based on multiple of the fundamental period. The result is shown with multiple plots.
To do:
1. Plot the key number relate to the faults.
2. Configure the figure size to better view the plots.
3. Save the plots in .jpg