Issue #10: Fix the fault state and fault mode.
Fault state and fault mode plots shows noise in the steady state reagion before and after the fault.
The error is due to the normal RMS current used to verify the states.
Enhancement: add plots of the normal RMS currents.
To do:
Apply the algorithm to detect the faulty switch;
Apply the algorithm to each fault simulation.