Issue#10: Add a manager of simulation data and plots.
Add the class AnalysisManager in the analysisManager.py to manage the data processed in the ConditionMonitoring class. While ConditionMonitoring class serves to process intervals of signal, AnalysisManager serves to slice the simulation data into intervals to feed the ConditionMonitoring class, further processing and give a plot support to the data extracted from the simulations. The core processing is developed in the ConditionMonitoring class.
Add a file to analysis all the single switch fault simulationAnalysis.py
To do:Enhance the plot of the data to be easier to plot any data the user want, with legends and axis labels.