Issue #10: Undo the last commit.
All the data from the faults in the .csv files are summing up to 5GB and the project need to be realocated in my computer as well as in Github because a single file exceeded 100MB, which is more than Github's limit.
To do:
Modify the way the data is read.
Find a new place to store the .csv file (cloud or external drive).